remove PermissionManager class & add permissions request code in MapFragment
